refactor: change browser.storage.local structure
This refactor aims to make the data a bit more organised so I can later
add the opt-out by default feature to displaying last seen grades
without having a headache.

`browser.storage.local`'s initial structure looked something like this:

```json
{
    "user_data": {
        "most_recent_user": "name",
        "user": {
            "stuff": "goes here",
        }
    }
}
```

I altered it to look like this:

```json
{
    "data": {
        "opted_in": true,
        "showExtensionInfo": true,
        "most_recent_user": "name",
        "user_data": {
            "user": {
                "stuff": "goes here",
            }
        }
    }
}
```

Originally, `user_data` was the main data stored. I changed it to be a
single part of the data that is stored. `most_recent_user` was inside of
`user_data`, but I felt it didn't fit, so I moved it outside into
`data`.

`opted_in` is the variable name I replaced for `save_last_grades`.
`save_last_grades` was originally accessed/created by doing:
```js
browser.storage.local.get({"save_last_grades": true});
```

When searching through the local storage by supplying an object as the
function argument, the fields in the object are used as default values.
So, despite `save_last_grades` not actually being a part of the stored
data, it was returned as part of the data. Changing this to search
regularly (with the name of the object as the key in the storage) fixes
this, as well as adding the field in the instantiation of `data`.

"showExtensionInfo" is some "legacy" (I'm using this tentatively) stuff
Gary added that I'm too lazy to figure out.

Signed-off-by: Lucas Sta Maria <lucas.stamaria@gmail.com>
Signed-off-by: Advay Ratan <advayratan@gmail.com>