fix: make saved data track if the value has been changed
`opted_in` and `showExtensionInfo`, instead of storing a boolean value,
now store an object consisting of a `value` and `changed` keys. `value`
indicates which value the user wants for the given option, and `changed`
indicates whether the value has been changed from its default value.
Adding the `changed` key allows for users to have their options kept the
same if the default value is changed.

Signed-off-by: Lucas Sta Maria <lucas.stamaria@gmail.com>