fix: check if local data is missing config data
Added a function `getLocalConfig` to retrieve everything in
browser.storage.local. According to MDN, passing a `null` or `undefined`
value retrieves the entire storage contents, so that is what I did.
Wrapped in a function for better readability and understanding.

In `saveGradesLocally`, I used the `getLocalConfig` function to get the
existing config. Then, checks are performed to see if the config values
exist, and to add them if they don't.

Before, `user_data` also erased previously stored user data. By
attempting to retrieve it from the local storage first, this prevents
that from happening.

Signed-off-by: Lucas Sta Maria <lucas.stamaria@gmail.com>